# Rewrite clean-nginx-conf.sh in Go to speed up admission webhook