## What this PR does / why we need it:
Current implementation of validating web hook may be too slow in some
circumstances because it use Shell-script to tidy generated full nginx config.
On my dev-cluster with lots of namespaces/ingresses this sometimes lead too (in apiserver):
```
Error: Internal error occurred: failed calling webhook "validate.nginx.ingress.kubernetes.io": Post https://nginx-internal-grpc-ingress-nginx-controller-admission.ingress-nginx.svc:443/networking/v1beta1/ingresses?timeout=30s: context deadline exceeded
``` 

web hook: https://github.com/kubernetes/ingress-nginx/blob/f5cfd5730c4b296c87fbc531c83a6e4f33483b75/internal/admission/controller/main.go#L57
call to shell-script: https://github.com/kubernetes/ingress-nginx/blob/f5cfd5730c4b296c87fbc531c83a6e4f33483b75/internal/ingress/controller/template/template.go#L113
shell-script: https://github.com/kubernetes/ingress-nginx/blob/f5cfd5730c4b296c87fbc531c83a6e4f33483b75/rootfs/ingress-controller/clean-nginx-conf.sh

I rewrite tidy-code in Golang in this PR to speed it up.

My kubernetes in AWS:
```
bash-5.0$ time /ingress-controller/clean-nginx-conf.sh < nginx-big.conf > result.old
real	0m3.696s
user	0m4.519s
sys	0m0.868s

bash-5.0$ time ./indent < nginx-big.conf > result.new

real	0m0.127s
user	0m0.120s
sys	0m0.008s

bash-5.0$ diff -u result.old result.new
--- result.old	2021-04-25 14:46:57.549149609 +0000
+++ result.new	2021-04-25 14:47:21.885195144 +0000
@@ -44,7 +44,7 @@
 			use_proxy_protocol = false,
 			is_ssl_passthrough_enabled = false,
 			http_redirect_code = 308,
-		listen_ports = { ssl_proxy = "442", https = "443" },
+			listen_ports = { ssl_proxy = "442", https = "443" },

 			hsts = true,
 			hsts_max_age = 15724800,
@@ -89,7 +89,7 @@
 		plugins = res
 		end
 		-- load all plugins that'll be used here
-	plugins.init({  })
+		plugins.init({  })
 	}

 	init_worker_by_lua_block {

bash-5.0$
```

On localhost (not so impressive):
```
17:57:23 cgorbit@cgorbit:[0]~/go/src/k8s.io/ingress-nginx$ time rootfs/ingress-controller/clean-nginx-conf.sh < ~/nginx-big.conf > result.old

real	0m0,328s
user	0m0,730s
sys	0m0,074s
17:57:29 cgorbit@cgorbit:[0]~/go/src/k8s.io/ingress-nginx$ time ./indent < ~/nginx-big.conf > result.new

real	0m0,132s
user	0m0,116s
sys	0m0,017s
17:57:31 cgorbit@cgorbit:[0]~/go/src/k8s.io/ingress-nginx$ diff -u result.old result.new
--- result.old	2021-04-25 17:57:24.120927547 +0300
+++ result.new	2021-04-25 17:57:31.020945713 +0300
@@ -44,7 +44,7 @@
 			use_proxy_protocol = false,
 			is_ssl_passthrough_enabled = false,
 			http_redirect_code = 308,
-		listen_ports = { ssl_proxy = "442", https = "443" },
+			listen_ports = { ssl_proxy = "442", https = "443" },

 			hsts = true,
 			hsts_max_age = 15724800,
@@ -89,7 +89,7 @@
 		plugins = res
 		end
 		-- load all plugins that'll be used here
-	plugins.init({  })
+		plugins.init({  })
 	}

 	init_worker_by_lua_block {
```

## Types of changes
- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing functionality to change)

This is performance improvement.

## Which issue/s this PR fixes
No related issues

## How Has This Been Tested?
1. I compare result of original clean-nginx-conf.sh and my new code on 11MB nginx.conf
   The only diff is: https://gist.github.com/cgorbit/c0c9d6442dc18d44113d685e2e6bce27
   Which is improvement, I think.
2. make test: https://gist.github.com/cgorbit/bd302d9fdd55a93d509427981f9ee490
3. make lua-test
```
$ make lua-test
nginx: [alert] could not open error log file: open() "/var/log/nginx/error.log" failed (13: Permission denied)
●●●●●●●●●●●●●●●●●●●●●●●●●●●●●●●●●●●●●●●●●●●●●●●●●●●●●●●●●●●●●●●●●●●●●●●●●●●●●●●●●●●●●●●●●●●●●●●●●●●●●●●●●●●●●●●●●●●●●●●●●●●●●●●●●●●●●●●●●●●●●●●●●●●●●●●●●●●●●●●●●●●
163 successes / 0 failures / 0 errors / 0 pending : 0.088464 seconds
```
4. make kind-e2e-test
https://gist.github.com/cgorbit/b6f2db83275c8c996f5bb5b4de668820

## Checklist:
- [ ] My change requires a change to the documentation.
- [ ] I have updated the documentation accordingly.
- [x] I've read the [CONTRIBUTION](https://github.com/kubernetes/ingress-nginx/blob/master/CONTRIBUTING.md) guide
- [ ] I have added tests to cover my changes.
- [x] All new and existing tests passed.