@cgorbit 2021-04-27
Sorry if I misunderstood instruction and make early assign