@paalkr 2021-06-02
@andersosthus compiled a custom build, for internal testing, with this fix on top of 0.46. In a K8s 1.18 cluster with approx 900 complex ingress objects (several hundred hosts, snippets, regex, certs etc) validation using the bash based code did not succeed within the 30 seconds limit. Now with this Go based implementation, validation takes around 4-7 seconds. So I would say this is a huge improvement. The previous implementation did not work for us due to the timeout.