@paalkr 2021-06-15
Any news on when this fix will be merged? We have been running with this in production for two weeks as of now, it really is a game changes for clusters with many ingress objects and complex configurations.