@rikatz 2021-06-15
Hi there,

I've been a bit busy with #7156 and this is why I'm taking some time to answer.

As soon as I finish that one I'm really willing to look at this PR, sounds really promising.

Thanks