@rikatz 2021-06-27
Hey @cgorbit yes I would like that we deal with the error instead of panic (as my previous comment) just to make sure that some change into future go version wont break things or cause panic in a future (even if today we know it always returns nil).

Also, this is a nit but if possible please document the behavior of the package/what exactly cleanConf is doing (tabulating, removing characters, etc). You can do this on a followup/later PR as well, but would be great to have this.

As soon as you think this is ready, ping me on Slack (my email is pretty messy...) and I will have a new look into this and hopefully merge. So we can also cherry pick this to the branch release-v1beta1 to be available in past ingress version.