@rikatz 2021-06-28
/label tide/merge-method-squash
@cgorbit Thanks for the changes :) 
I'll wait until the CI passes to merge this.

As this is a pretty important change, after this gets merged, would you mind opening a PR/cherrypick to branch release-v1beta1 to be also inside the old v0.X releases?

Thanks